requirements2.txt
# SPDX-FileCopyrightText: 2021 Forschungs- und Entwicklungszentrum Fachhochschule Kiel GmbH
# 
# SPDX-License-Identifier: CC0-1.0

python-can==3.3.4
cantools==36.2.0
pytest==5.4.3
pytest-mock==3.1.1
pytest-timeout==1.4.2
pytest-console-scripts==1.1.0
pyserial==3.5b0
PyYAML==5.3.1
scipy==1.6.1
pydantic==1.8.2
tabulate==0.8.9
sphinx_rtd_theme
# git+https://github.com/maltekliemann/controllino@0.1.1
# git+https://github.com/ppannuto/python-saleae@877178f67618cdb3355054eae3666d99dce36aaa
# git+https://github.com/maltekliemann/rogue@0.1.7
